requirements.txt
gymnasium>=0.29
matplotlib>=3.8
numpy>=1.26
pandas>=2.2
python-dotenv>=1.0
scikit-learn>=1.4
streamlit>=1.32
tensorboard>=2.15
torch>=2.2
tqdm>=4.66
